requirements.txt
pandas 
numpy
mysql-connector-python
